+++ lib/Control/Monad/Dep/Class.hs
@@ -65,6 +65,8 @@ module Control.Monad.Dep.Class
 
     -- * Lifting effects from dependencies
     LiftDep (..),
+    -- * Helpers
+    withEnv
   )
 where
 
@@ -136,3 +138,10 @@ type MonadDep ::
 type family MonadDep dependencies d e m where
   MonadDep '[] d e m = (LiftDep d m, MonadReader e m)
   MonadDep (dependency ': dependencies) d e m = (dependency d e, MonadDep dependencies d e m)
+
+-- | Avoids repeated calls to 'liftD' when all the effects in a function come from the environment.
+withEnv :: forall d e m r. (LiftDep d m, MonadReader e m) => (e -> d r) -> m r
+withEnv f = do
+  e <- ask
+  liftD (f e)
+