+++ lib/Control/Monad/Dep/Has.hs
@@ -11,25 +11,30 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ViewPatterns #-}
 
--- | This module provides a generic \"Has\" class favoring a style in which the
--- components of the environment come wrapped in records or newtypes, instead
--- of being bare functions.
+-- | This module provides a general-purpose \"Has\" class favoring a style in
+-- which the components of the environment come wrapped in records or newtypes,
+-- instead of being bare functions.
 --
 -- >>> :{
 --  type Logger :: (Type -> Type) -> Type
 --  newtype Logger d = Logger {log :: String -> d ()} deriving Generic
 --  instance Dep Logger where
 --    type DefaultFieldName Logger = "logger"
+--  --
 --  data Repository d = Repository
 --    { select :: String -> d [Int],
 --      insert :: [Int] -> d ()
 --    } deriving Generic
 --  instance Dep Repository where
 --    type DefaultFieldName Repository = "repository"
+--  --
 --  newtype Controller d = Controller {serve :: Int -> d String} deriving Generic
 --  instance Dep Controller where
 --    type DefaultFieldName Controller = "controller"
+--  --
 --  type Env :: (Type -> Type) -> Type
 --  data Env m = Env
 --    { logger :: Logger m,
@@ -39,13 +44,13 @@
 --  instance Has Logger m (Env m)
 --  instance Has Repository m (Env m)
 --  instance Has Controller m (Env m)
---  mkController :: forall d e m. MonadDep [Has Logger, Has Repository] d e m => Controller m
+--  --
+--  mkController :: MonadDep [Has Logger, Has Repository] d e m => Controller m
 --  mkController =
---    Controller \url -> do
---      e <- ask
---      liftD $ log (dep e) "I'm going to insert in the db!"
---      liftD $ select (dep e) "select * from ..."
---      liftD $ insert (dep e) [1, 2, 3, 4]
+--    Controller \url -> useSelf \self -> do
+--      self log "I'm going to insert in the db!"
+--      self select "select * from ..."
+--      self insert [1, 2, 3, 4]
 --      return "view"
 -- :}
 --
@@ -53,16 +58,17 @@
 -- \"dep-t-advice\" can facilitate working with this style of components.
 --
 module Control.Monad.Dep.Has (
-        -- * A generic \"Has\"
-        Has (..), 
+        -- * A general-purpose Has
+        Has (..) 
         -- * Component defaults
-        Dep (..)
+    ,   Dep (..)
+        -- * self
+    ,   useSelf
     ) where
 
-import Data.Kind
-import GHC.Records
-import GHC.TypeLits
-import Data.Coerce
+import DI
+import Control.Monad.Dep.Class
+import Control.Monad.Reader.Class
 
 -- $setup
 --
@@ -75,58 +81,25 @@ import Data.Coerce
 -- >>> :set -XFunctionalDependencies
 -- >>> :set -XFlexibleContexts
 -- >>> :set -XDataKinds
--- >>> :set -XRankNTypes
 -- >>> :set -XBlockArguments
 -- >>> :set -XFlexibleInstances
 -- >>> :set -XTypeFamilies
 -- >>> :set -XDeriveGeneric
+-- >>> :set -XViewPatterns
+-- >>> import Data.Kind
 -- >>> import Control.Monad.Dep
 -- >>> import Rank2 qualified
 -- >>> import Rank2.TH qualified
 -- >>> import GHC.Generics (Generic)
 --
 
-
--- | A generic \"Has\" class. When partially applied to a parametrizable
--- record-of-functions @r_@, produces a 2-place constraint that can be later
--- used with "Control.Monad.Dep.Class".
-type Has :: ((Type -> Type) -> Type) -> (Type -> Type) -> Type -> Constraint
-class Has r_ d e | e -> d where
-  -- |  Given an environment @e@, produce a record-of-functions parameterized by the environment's effect monad @d@.
-  --
-  -- The hope is that using a selector function on the resulting record will
-  -- determine its type without the need for type annotations.
-  --
-  -- (This will likely not play well with RecordDotSyntax. See also <https://chrisdone.com/posts/import-aliases-field-names/ this trick>.)
-  dep :: e -> r_ d
-  default dep :: (Dep r_, HasField (DefaultFieldName r_) e u, Coercible u (r_ d)) => e -> r_ d
-  dep e = coerce . getField @(DefaultFieldName r_) $ e
-
--- | Parametrizable records-of-functions can be given an instance of this
--- typeclass to specify the default field name 'Has' expects for the component
--- in the environment record.
+-- asSelf :: forall env m . env -> forall r_ x. Has r_ m env => (r_ m -> x) -> x
+-- asSelf env = \f -> f (dep env)
 --
--- This allows defining 'Has' instances with empty bodies, thanks to
--- @DefaultSignatures@.
-type Dep :: ((Type -> Type) -> Type) -> Constraint
-class Dep r_ where
-  -- The Char kind would be useful here, to lowercase the first letter of the
-  -- k type and use it as the default preferred field name.
-  type DefaultFieldName r_ :: Symbol
 
--- -- Doesn't make much sense to have this, we already have Has!
--- type Sub :: ((Type -> Type) -> Type) -> ((Type -> Type) -> Type) -> Constraint
--- class Sub sub super
--- type SubWrapper :: ((Type -> Type) -> Type) -> ((Type -> Type) -> Type) -> (Type -> Type) -> Type -> Type
--- data SubWrapper sub super d e = SubWrapper e
-
--- type Nested :: ((Type -> Type) -> Type) -> ((Type -> Type) -> Type) -> (Type -> Type) -> Type -> Type
--- data Nested sub super d e = Nested e
--- 
--- instance (Has sub d e, Has super d (sub d)) => Has super d (Nested sub super d e) where
---     dep (Nested e) = dep @super (dep @sub e)
-
--- Possible example
--- instance Has ReadRef IO (Env IO) via (Nested Ref ReadRef IO (Env IO))
-
+-- | Avoids repeated calls to 'liftD' when all the effects in a function come from the environment.
+useSelf :: forall d env m y . (LiftDep d m, MonadReader env m) => ((forall r_ x . Has r_ d env => (r_ d -> x) -> x) -> d y) -> m y
+useSelf needsSelf = do
+  (asSelf -> self) <- ask @env
+  liftD (needsSelf self)