+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.4.4.0
 
-* added Control.Monad.Dep.Has
+* added Control.Monad.Dep.Has, a generic "Has" typeclass which favors a style in which
+  the components come wrapped in records or newtypes.
+
+* added "withEnv" to Control.Monad.Dep.Class.
 
 ## 0.4.0.0