+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for dep-t
 
+## 0.4.5.0
+
+* the definitions from Control.Monad.Dep.Has can be useful in scenarios
+  unrelated to DepT and MonadDep. They have been moved to a separate package
+  "dani-di" and re-exported from there.
+
+* added "userEnv" to Control.Monad.Dep.Has
+
 ## 0.4.4.0
 
 * added Control.Monad.Dep.Has, a generic "Has" typeclass which favors a style in which